refactor: 클래스 파일 구조 변경 및 CommentsUtils 추가
