+++ src/main/java/com/single/springboard/service/comments/CommentsService.java
@@ -1,4 +1,4 @@
-package com.single.springboard.service.posts;
+package com.single.springboard.service.comments;
 
 import com.single.springboard.domain.comments.Comments;
 import com.single.springboard.domain.comments.CommentsRepository;
